Move all components of card to their own folder
