+++ src/main/java/jalau/at18/kingoftokyo/view/GameFrame.java
@@ -5,6 +5,8 @@ import jalau.at18.kingoftokyo.controller.GameFrameController;
 import jalau.at18.kingoftokyo.view.card.CardPanel;
 import jalau.at18.kingoftokyo.*;
 
+import jalau.at18.kingoftokyo.view.card.CardsGroupPanel;
+
 public class GameFrame extends JFrame {
     private static final int DEFAULT_POS = 30;
     private static final int DEFAULT_WIDTH = 1200;
