+++ src/main/java/jalau/at18/kingoftokyo/view/card/CardsGroupPanel.java
@@ -1,4 +1,4 @@
-package jalau.at18.kingoftokyo.view;
+package jalau.at18.kingoftokyo.view.card;
 
 import javax.swing.Box;
 import javax.swing.BoxLayout;
@@ -19,10 +19,10 @@ public class CardsGroupPanel extends JPanel {
     public CardsGroupPanel() {
         setBounds(CARD_POSX, CARD_POSY, CARD_WIDTH, CARD_HEIGHT);
         setLayout(new BoxLayout(this, BoxLayout.LINE_AXIS));
-        add(new CardPanel(0, new CompleteCard(DiscardCards.ENERGIZE)));
+        add(new CardPanel(new CompleteCard(DiscardCards.ENERGIZE)));
         add(Box.createRigidArea(new Dimension(CARDS_SEPARATION, 0)));
-        add(new CardPanel(1, new CompleteCard(DiscardCards.COMMUTER_TRAIN)));
+        add(new CardPanel(new CompleteCard(DiscardCards.COMMUTER_TRAIN)));
         add(Box.createRigidArea(new Dimension(CARDS_SEPARATION, 0)));
-        add(new CardPanel(2, new CompleteCard(DiscardCards.APARTMENT_BULIDING)));
+        add(new CardPanel(new CompleteCard(DiscardCards.APARTMENT_BULIDING)));
     }
 }
