+++ src/test/java/jalau/at18/kingoftokyo/view/CardPanelTest.java
@@ -11,17 +11,17 @@ import org.junit.Test;
 public class CardPanelTest {
     @Test
     public void shouldShowACard() {
-        CardPanel cardPanel1 = new CardPanel(0, new CompleteCard(DiscardCards.ENERGIZE));
+        CardPanel cardPanel1 = new CardPanel(new CompleteCard(DiscardCards.ENERGIZE));
         assertEquals(6, cardPanel1.getComponentCount());
         assertEquals(150, cardPanel1.getBounds().getWidth(), 0);
         assertEquals(220, cardPanel1.getBounds().getHeight(), 0);
         assertTrue(cardPanel1.isVisible());
 
-        CardPanel cardPanel2 = new CardPanel(1, new CompleteCard(DiscardCards.ENERGIZE));
+        CardPanel cardPanel2 = new CardPanel(new CompleteCard(DiscardCards.ENERGIZE));
         assertEquals(6, cardPanel2.getComponentCount());
         assertTrue(cardPanel2.isVisible());
 
-        CardPanel cardPanel3 = new CardPanel(2, new CompleteCard(DiscardCards.ENERGIZE));
+        CardPanel cardPanel3 = new CardPanel(new CompleteCard(DiscardCards.ENERGIZE));
         assertEquals(6, cardPanel3.getComponentCount());
         assertTrue(cardPanel3.isVisible());
     }
